Drop the elisp variable "sr-speedbar-window".
Tracking the window displaying the speedbar buffer all the time leads to
inconsistencies, especially when the configuration of windows changes.
Therefore, the current mechanism,
which relied on a parameter and many hooks/advices to keep track of
the sr-speedbar window was replaced by
a "dynamic" method (i.e., a function) which retrieves the sr-speedbar window
when needed.